requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.7
Django==4.2.7
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
filelock==3.13.1
fonttools==4.39.4
idna==3.6
kiwisolver==1.4.4
lark-parser==0.12.0
matplotlib==3.7.1
mysqlclient==2.2.3
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
pipenv==2023.11.15
platformdirs==3.11.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.0
virtualenv==20.24.6
